+++ assets/javascripts/kitten/components/form/location-input/index.js
@@ -26,7 +26,7 @@ const StyledLocationInput = styled.div`
     width: 100%;
     height: ${pxToRem(50)};
     background: ${COLORS.background1};
-    border: ${pxToRem(2)} solid ${COLORS.line1};
+    border: var(--border-width) solid var(--color-grey-400);
     border-radius: 0;
     color: ${COLORS.font1};
 
@@ -38,7 +38,7 @@ const StyledLocationInput = styled.div`
 
     &:focus {
       color: ${COLORS.font1};
-      border-color: ${COLORS.line2};
+      border-color: var(--color-grey-500);
     }
 
     &:disabled {
@@ -63,7 +63,7 @@ const StyledLocationInput = styled.div`
   .k-LocationInput__autocomplete {
     box-sizing: border-box;
     background-color: ${COLORS.background1};
-    border: ${pxToRem(2)} solid ${COLORS.line1};
+    border: var(--border-width) solid var(--color-grey-400);
     border-top: 0;
     position: absolute;
     z-index: 1000;