+++ assets/javascripts/kitten/components/molecules/boxes/link-box/styles.js
@@ -21,7 +21,7 @@ export const StyledLinkBox = styled.a`
     box-sizing: border-box;
     color: ${COLORS.font1};
     background-color: var(--LinkBox-background-color);
-    border: ${pxToRem(2)} solid ${COLORS.line1};
+    border: var(--border-width) solid var(--color-grey-400);
     border-radius: ${pxToRem(4)};
     transition: background-color 0.2s ease;
 
@@ -74,11 +74,11 @@ export const StyledLinkBox = styled.a`
     padding-bottom: ${pxToRem(7)};
     border-radius: ${pxToRem(8)};
     background-color: var(--LinkBox-background-color);
-    border: ${pxToRem(2)} solid var(--LinkBox-background-color);
+    border: var(--border-width) solid var(--LinkBox-background-color);
 
     &:hover {
       background-color: var(--LinkBox-background-color);
-      border: ${pxToRem(2)} solid var(--LinkBox-background-color);
+      border: var(--border-width) solid var(--LinkBox-background-color);
 
       .k-LinkBox__arrow {
         transform: translate(${pxToRem(5)}, ${pxToRem(0)});
@@ -87,7 +87,7 @@ export const StyledLinkBox = styled.a`
 
     &:active {
       background-color: var(--LinkBox-background-color);
-      border: ${pxToRem(2)} solid var(--LinkBox-background-color);
+      border: var(--border-width) solid var(--LinkBox-background-color);
     }
 
     .k-LinkBox__arrow {